requirements.txt
dash
dash-renderer
dash-core-components
dash-html-components
dash-table
dash-bootstrap-components
plotly
numpy
pandas
scipy
matplotlib
seaborn
requests
python-dotenv
ibis-framework[bigquery]
google-cloud-bigquery
# duckdb==0.9.2